services/metadata/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
psycopg2-binary==2.9.10
pydantic==2.9.2
httpx==0.27.2
nats-py==2.6.0
